Issue 1340: studentFeedbackSessionResults: view results for MCQ question
Update Issue 1340
Status: Delivered
For now I only display the choice chosen by the giver as text.

Is it better if students can see a compiled result (e.g. how many people choose a certain option) or is it fine as it is?

I assume the instructor's view should have a way to get the compiled result?