Issue 1368: Use inheritance for Feedback*Details class operations
Update Issue 1368
Status: ChangesRequested
I tried using inheritance for the Feedback*Details classes but I am currently stuck in two areas:
1) Creation of the Feedback*Details object
These takes variable number of arguments depending on the question type and I'm not sure hw to implement it as an abstract method as the abstract method wouldn't know what each sub-class needs. Probably have to pass in the whole Http request parameters?
2) Displaying new question creation form
The current implementation is designed to have all the forms on the page initially and then hidden/shown using javascript. This doesn't really lend well to the inheritance approach as all the different forms will have to be printed out anyway